Refactor : 버튼 컴포넌트로 합침, 불필요 컴포넌트 삭제
