+++ src/Pages/TodoPage/OtherCadetList/TodoCadetList.jsx
@@ -1,6 +1,5 @@
-import isToday from 'Utils/isToday';
-
-import CadetListCheckbox from './CadetListCheckbox';
+import { color, lineThrough } from 'Utils/TodoCss';
+import Checkbox from '@mui/material/Checkbox';
 
 import styled from 'styled-components';
 
@@ -14,7 +13,15 @@ const TodoCadetList = ({ item, date }) => {
         <TodoCadetListContainer>
           {item.todoDtoList.map((list, index) => (
             <TodoCadetListEntity key={index}>
-              <CadetListCheckbox list={list} />
+              <Checkbox
+                checked={list.titleCheck}
+                className="checkbox"
+                size="small"
+                sx={{
+                  '& .MuiSvgIcon-root': { fontSize: 15 },
+                }}
+                disabled
+              />
               <CadetListItem list={list} today={today} date={date}>
                 {list.title.replace(/ /g, '\u00a0')}
               </CadetListItem>
@@ -42,12 +49,14 @@ const WarningNoList = styled.div`
 const TodoCadetListContainer = styled.div``;
 const TodoCadetListEntity = styled.div`
   display: table;
+  .checkbox {
+    margin-left: -10px;
+    margin-bottom: -5px;
+  }
 `;
 const CadetListItem = styled.span`
-  color: ${props =>
-    !isToday(props.today, props.date) || props.list.titleCheck ? 'gray' : ''};
-  text-decoration-line: ${props =>
-    props.list.titleCheck ? 'line-through' : ''};
+  color: ${props => color(props, props.list.titleCheck)};
+  text-decoration-line: ${props => lineThrough(props.list.titleCheck)};
 `;
 
 export default TodoCadetList;
